moved wizard typings to new file
